requirements.txt
streamlit
crewai
python-dotenv
openai
langchain
langchain-openai
